Fracture on key as well as hash, interface update.
Now have the ability to fracture on a key created separate fracture
queues for each key. The queues will be kept in an LRU-cache and
disposed of if they fall out of use. The same queues are used for
hashing.

The presence of a `buckets` parameter means that the key extracted is
hashed and the hashed value is used to select the fractured queue.
Otherwise the key extracted is used directly.

Updated the interface to be identical to Turnstile Redux. Fracture
should be a drop in replacement for Turnstile and can be used with the
Turnstile Queue, Set and Check.