Returning entry and a future.
Now returning a user constructed entry and a future from `enqueue`. This
is because it is a common case to want to sometimes wait for work to
finish. Also, this future has a specific type, so we can use it to
indicate that the function you've called is fractured.

I'm finding that most background work occurs in the background with
occasional inquires as to its progress. In the case of the write-ahead
queue, we want to process writes in the background and on occasion we
want to sync those writes. We may as well build this into Fracture since
the construction of a single `Future` isn't that expensive if no one
waits on it.

By returning a `Fracture::Completion` your function can indicate to
other services that use fracture that they are going to depend upon the
completion of a function in a `Fracture` queue which may be the same
queue in which they reside, so they may choose to pause themselves to
allow the function its time in the queue.