Fix Pause synchronous capture.
Pulled Fracture concepts out of Strata. Didn't pull them out correctly.
Pause is supposed to be your synchronous time spent with the queue
contents for a given key. Pause was constructed with a copy of the
content and then returned from an `async` function. That is a race
because the contents of the queue can change before the `Promise`
returned from the function resolves.