Implement `CompletionSet`.
A class instead of the object we've been passing into `write` in Strata.
Created a null set so that I won't resist using this structure. A
completion set ought to make it clear to the caller that a method uses
Fracture so if taht method is called from Fracture the caller should use
`completed`.