Empty queue on error.
Empty the queued work on error. Pause futures are rejected. Futures
associated with queued work are rejected. In-process workers should be
able to complete. We let completions proceed so the worker can finish.

It is now important that workers release pauses in finally blocks to
allow the Fracture to work its way to zero.